interpretador

+++ grupo70/projJava/src/sortParbyValue.java
@@ -2,6 +2,6 @@ import java.util.Comparator;
 
 public class sortParbyValue implements Comparator<ParStringInt>{
     public int compare(ParStringInt normal, ParStringInt aux) {
-        return -(normal.getValue() - aux.getValue());
+        return (int)-(normal.getValue() - aux.getValue());
     }
 }
